hscope: Update telegram example
Now it takes into account the mountPoint. I thought that maybe I could
inject the mountPoint directly into the Lua environment and have it
available in on_broker_connected handlers. However, mountPoint is
available only through an rpc_call to the broker and that'll be
available only after the on_broker_connected handlers.