hscope: Rework status
`status` is now a separate node inside a tester node. This makes it
possible to have multiple values inside the HscopeNode each with their
own getter and chng signal. Having a directory called "status" is now
forbidden as it is a special directory.

The dashboard has also been reworked to adjust for the changes.