shvagent: Include device id in script dir
When multiple agents are running on the same pc, there is a race a race
condition when using `runScript`, on both of them at the same time with
the same script. shvagent always writes the contents of the script to
disk (if the user is not explicitly using hashes), so if both do it,
there's a race.

Now we include the device id in the script dir, so all agents use
different script dirs.