history: tests: Add explicit yield macros
In cases where I yield, I would like the message to be emitted AFTER I
go back into the event loop - I use a singleshot timer for that. The
reason is that a test message could result in the client emitting some
client messages. This means that the client emits directly inside my
coroutine, wihch is undesirable, because it interleaves the messages. A
drawback is that you have to explicitly state whether you want to yield,
but that's not such a big problem, I think. Basically, if you send a
message from the test suite and expect an immediate response, you should
use *_YIELD. If you do not expect a response, you shouldn't use the
*_YIELD variant.

The Mock class checks for this. Accepting another message, while you're
already handling one is forbidden.

Sometimes, I don't want to go to the event loop, especially if I don't
except a response immediately, for example, if the test doesn't end with
the client sending a message.