history: tests: Rework MockRpcConnection
... again. Sigh. The problem is this: suppose you get two requests from
the client, where the client doesn't wait for a response. You handle one
with EXPECT_REQUEST_YIELD and the second one with EXPECT_REQUEST.
However, since it's two messages, you can't be sure that the timer from
EXPECT_REQUEST_YIELD that calls doRespond resolves before you enter the
coroutine again. If it doesn't, that means that the already-responded-to
message will be still at the top of the message queue (because the timer
still didn't resolve). This is undesirable, because you don't care about
the already-responded-to message.

Solve this by doing doRespond and only use the timer for emitting of the
message.