history: Ignore events that come from leaf nodes we don't know
If there was a slave hp and some node we don't know about sent us an
event, we would assume that is it one of our leaf nodes. That's not the
all the time, as child nodes of a slave HP can still send events even
though they aren't actuallt logged anywhere.