historyprovider3: Trim last milliseconds only for files we create
Which are the files we create from getLog. We mustn't edit the file we
download by syncing, because they might not be consistent upon writing
by us.

getLog files are still trimmed, because we create those. We discussed
the last-ms algorithm and figured that the situation doesn't happen
often (if ever), so we don't care. The getLog situation is still the
same, so we still need the algo.